+++ src/probnum/quad/solvers/policies/__init__.py
@@ -1,1 +1,2 @@
-from ._policy import Policy, RandomPolicy
+from ._policy import Policy
+from ._random_policy import RandomPolicy
